+++ frontend/src/components/Header/constants.ts
@@ -1,0 +1,8 @@
+import { ROUTES } from "src/common/constants/routes";
+
+export const FOOTER_OPTIONS = [
+  { id: 1, name: "Documentation", value: ROUTES.WMG_DOCS },
+  { id: 2, name: "Contact", value: "mailto:cellxgene@chanzuckerberg.com" },
+  { id: 3, name: "Terms of Service", value: ROUTES.TOS },
+  { id: 4, name: "Privacy Policy", value: ROUTES.PRIVACY },
+];
